add loger for service impl file

+++ src/main/java/com/example/javaecommerce/services/impl/OrderDetailServiceImpl.java
@@ -11,6 +11,8 @@ import com.example.javaecommerce.repository.OrderDetailRepository;
 import com.example.javaecommerce.services.OrderDetailService;
 
 import lombok.RequiredArgsConstructor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -22,38 +24,76 @@ public class OrderDetailServiceImpl implements OrderDetailService {
 
     private final OrderDetailMapper orderDetailMapper;
 
+    private static final Logger logger = LoggerFactory.getLogger(OrderDetailServiceImpl.class);
+
     @Override
     public List<OrderDetailResponse> getAllOrderDetails() {
-        List<OrderDetailEntity> orderDetailEntities = orderDetailRepository.findAll();
-        return orderDetailMapper.toListOrderDetailResponse(orderDetailEntities);
+        try {
+            List<OrderDetailEntity> orderDetailEntities = orderDetailRepository.findAll();
+            logger.info("Get all order details with pagination successfully");
+            return orderDetailMapper.toListOrderDetailResponse(orderDetailEntities);
+        } catch (Exception e) {
+            logger.info("Failed to get all order details with pagination", e);
+            throw new RuntimeException("Failed to get all order details with pagination");
+        }
     }
 
     @Override
     public OrderDetailResponse addOrderDetail(final OrderDetailRequest orderDetailRequest) {
-        OrderDetailEntity orderDetailEntity = Converter.toModel(orderDetailRequest, OrderDetailEntity.class);
-        orderDetailRepository.save(orderDetailEntity);
-        return orderDetailMapper.toOrderDetailResponse(orderDetailEntity);
+        try {
+            OrderDetailEntity orderDetailEntity = Converter.toModel(orderDetailRequest, OrderDetailEntity.class);
+            orderDetailRepository.save(orderDetailEntity);
+            logger.info("Create order successfully!");
+            return orderDetailMapper.toOrderDetailResponse(orderDetailEntity);
+        } catch (Exception e) {
+            logger.info("Failed to create order");
+            throw new RuntimeException("Failed to create order", e);
+        }
     }
 
     @Override
     public OrderDetailResponse getOrderDetailById(final Long orderDetailId) {
-        OrderDetailEntity orderDetailEntity = orderDetailRepository.findById(orderDetailId)
-                .orElseThrow(() -> new EcommerceRunTimeException(ErrorCode.ID_NOT_FOUND));
-        return orderDetailMapper.toOrderDetailResponse(orderDetailEntity);
+        try {
+            OrderDetailEntity orderDetailEntity = orderDetailRepository.findById(orderDetailId)
+                    .orElseThrow(() -> new EcommerceRunTimeException(ErrorCode.ID_NOT_FOUND));
+            logger.info("Get order detail successfully by id {}", orderDetailId);
+            return orderDetailMapper.toOrderDetailResponse(orderDetailEntity);
+        } catch (Exception e) {
+            logger.info("Failed to get order detail by id", e);
+            throw new RuntimeException("Failed to get order detail by id", e);
+        }
     }
 
     @Override
     public void deleteOrderDetail(final Long orderDetailId) {
-        orderDetailRepository.deleteById(orderDetailId);
+        try {
+            var orderDetails = orderDetailRepository
+                    .findById(orderDetailId)
+                    .orElseThrow(() -> new EcommerceRunTimeException(ErrorCode.ID_NOT_FOUND));
+            if (orderDetails != null) {
+                logger.info("order detail is found by id {}", orderDetailId);
+                orderDetailRepository.deleteById(orderDetailId);
+                logger.info("delete order detail successfully by id {}", orderDetailId);
+            }
+        } catch (Exception e) {
+            logger.info("Failed to get order detail by id", e);
+            throw new RuntimeException("Failed to get order detail by id", e);
+        }
     }
 
     @Override
     public OrderDetailResponse updateOrderDetail(final OrderDetailRequest orderDetailRequest, final Long id) {
-        OrderDetailEntity orderDetailEntity = orderDetailRepository.findById(id)
-                .map(orderDetail -> {
-                    orderDetail.setTotal(orderDetailRequest.getTotal());
-                    return orderDetailRepository.save(orderDetail);
-                }).orElseThrow(() -> new EcommerceRunTimeException(ErrorCode.ID_NOT_FOUND));
-        return orderDetailMapper.toOrderDetailResponse(orderDetailEntity);
+        try {
+            OrderDetailEntity orderDetailEntity = orderDetailRepository.findById(id)
+                    .map(orderDetail -> {
+                        orderDetail.setTotal(orderDetailRequest.getTotal());
+                        return orderDetailRepository.save(orderDetail);
+                    }).orElseThrow(() -> new EcommerceRunTimeException(ErrorCode.ID_NOT_FOUND));
+            logger.info("Update order details successfully by id {} ", id);
+            return orderDetailMapper.toOrderDetailResponse(orderDetailEntity);
+        } catch (Exception e) {
+            logger.info("Failed to update order details", e);
+            throw new RuntimeException("Failed to update order details");
+        }
     }
 }
